BackEnd/cricBeeproject/.history/requirements_20251122211640.txt
aiosmtplib==5.0.0
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
bcrypt==4.0.1
billiard==4.2.2
boto3==1.41.2
botocore==1.41.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
cron_descriptor==2.0.6
cryptography==46.0.3
distlib==0.3.9
dj-rest-auth==7.0.1
dnspython==2.8.0
drf-spectacular==0.28.0
ecdsa==0.19.1
email-validator==2.3.0
Faker==33.3.1
fastapi==0.121.3
filelock==3.17.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
platformdirs==4.3.6
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.15.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.29.2
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
